feat(webui): add /settings route with full-page settings view
Extract shared settings panels into a reusable SettingsContent component
used by both the existing SettingsModal dialog and the new full-page
Settings route. The CommandPalette Settings action now navigates to
/settings instead of opening the modal.

- Add /settings to appRoutes and cloudflareSpaRedirectRules
- Create SettingsContent shared component with category sidebar + panels
- Create SettingsPage with full-page layout (MenuBar, SidebarIcons)
- Refactor SettingsModal to use SettingsContent
- Update CommandPalette to navigate to /settings
- Update public/_redirects for Cloudflare SPA fallback
- Update test to expect navigation instead of modal open

Closes gptme/gptme#2799
